Added the move Reflect Type with its generation 5 effect.

+++ include/pokemon.h
@@ -238,7 +238,7 @@ struct SpeciesInfo
 
 struct BattleMove
 {
-    u8 effect;
+    u16 effect;
     u8 power;
     u8 type;
     u8 category;
